+++ knowthenics-data/src/main/java/ru/dpopkov/knowthenics/model/BaseEntity.java
@@ -15,4 +15,8 @@ public class BaseEntity implements Serializable {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
+
+    public boolean isNew() {
+        return id == null;
+    }
 }