+++ knowthenics-data/src/main/java/ru/dpopkov/knowthenics/model/Category.java
@@ -21,4 +21,9 @@ public class Category extends BaseEntity {
         this.name = name;
         this.description = description;
     }
+
+    @Override
+    public String toString() {
+        return name;
+    }
 }