+++ History.md
@@ -1,5 +1,8 @@
 # History
 
+#### 2021-01-17
+* Add ability to add or update a Question. Closes #37.
+
 #### 2021-01-16
 * Display Question Record. Closes #34.
 * Add in the Question Details page link to edit Answer, link to view Answer Details page. Closes #35.